packageCreate OpenTapIgnore: Fix
Using AssemblyFinder for locating assemblies in a more safe way than what PluginSearcher did previously. This also ensures that it respects OpenTapIgnore.
